Avoid some division by zero Set blast_range and blast_force to uint type (fix values loaded from xml files)

+++ trunk/src/weapon/weapon_cfg.h
@@ -52,8 +52,8 @@ class ExplosiveWeaponConfig : public WeaponConfig
 public:
   uint timeout;
   bool allow_change_timeout;
-  double blast_range ;
-  double blast_force ;  
+  uint blast_range ;
+  uint blast_force ;  
   uint explosion_range;
   uint particle_range;
 public:
